Move handle_request into a separate source file
