+++ libraries/psibase_http/http.cpp
@@ -6,12 +6,10 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include "psibase/http.hpp"
-#include "psibase/Socket.hpp"
-#include "psibase/TransactionContext.hpp"
 #include "psibase/log.hpp"
-#include "psibase/serviceEntry.hpp"
 
-#include "websocket_log_session.hpp"
+#include "http_session_base.hpp"
+#include "server_state.hpp"
 
 #include <boost/asio/bind_executor.hpp>
 #include <boost/asio/io_service.hpp>
@@ -28,15 +26,10 @@
 #include <boost/make_unique.hpp>
 #include <boost/optional.hpp>
 #include <boost/signals2/signal.hpp>
-#include <boost/type_erasure/is_empty.hpp>
-
-#include <psio/finally.hpp>
-#include <psio/to_json.hpp>
 
 #include <algorithm>
 #include <charconv>
 #include <cstdlib>
-#include <fstream>
 #include <functional>
 #include <iostream>
 #include <memory>
@@ -59,10 +52,6 @@ using std::chrono::steady_clock;  // To create explicit timer
 
 namespace psibase::http
 {
-   struct websocket_upgrade
-   {
-   };
-
    // Report a failure
    static void fail(psibase::loggers::common_logger& logger, beast::error_code ec, const char* what)
    {
@@ -101,19 +90,16 @@ namespace psibase::http
       std::function<void()> callback;
    };
 
-   struct server_impl
+   struct server_impl : server_state
    {
-      net::io_service                          ioc;
-      std::shared_ptr<const http::http_config> http_config = {};
-      std::shared_ptr<psibase::SharedState>    sharedState = {};
-      std::vector<std::thread>                 threads     = {};
-      using signal_type                                    = boost::signals2::signal<void(bool)>;
-      signal_type      shutdown_connections;
+      net::io_service          ioc;
+      std::vector<std::thread> threads = {};
+      using signal_type                = boost::signals2::signal<void(bool)>;
       shutdown_tracker thread_count;
 
       server_impl(const std::shared_ptr<const http::http_config>& http_config,
                   const std::shared_ptr<psibase::SharedState>&    sharedState)
-          : http_config{http_config}, sharedState{sharedState}
+          : server_state{http_config, sharedState}
       {
       }
 
@@ -131,21 +117,7 @@ namespace psibase::http
          sharedState.reset();
       }
 
-      template <typename T>
-      void register_connection(const std::shared_ptr<T>& ptr)
-      {
-         shutdown_connections.connect(
-             signal_type::slot_type([weak = std::weak_ptr{ptr}](bool restart)
-                                    { T::close(weak.lock(), restart); })
-                 .track_foreign(std::weak_ptr{ptr}));
-      }
    };  // server_impl
-
-   class http_session_base;
-   template <class Body, class Allocator>
-   void handle_request(server_impl&                                           server,
-                       bhttp::request<Body, bhttp::basic_fields<Allocator>>&& req,
-                       http_session_base&                                     send);
 
    void server_service::async_close(bool restart, std::function<void()> callback)
    {
@@ -160,1479 +132,6 @@ namespace psibase::http
          impl->stop();
       }
    }
-
-   // This queue is used for HTTP pipelining.
-   class http_session_base : public std::enable_shared_from_this<http_session_base>
-   {
-     protected:
-      using message_type   = bhttp::response<bhttp::vector_body<char>>;
-      using request_type   = bhttp::request<bhttp::vector_body<char>>;
-      ~http_session_base() = default;
-
-      http_session_base(server_impl& server) : server(server)
-      {
-         static_assert(limit > 0, "queue limit must be positive");
-         items.reserve(limit);
-      }
-
-      enum
-      {
-         // Maximum number of responses we will queue
-         limit = 8
-      };
-
-      // The type-erased, saved work item
-      struct work
-      {
-         virtual ~work()           = default;
-         virtual void operator()() = 0;
-      };
-
-      std::vector<std::unique_ptr<work>> items;
-
-      void start_socket_timer()
-      {
-         auto self = shared_from_this();
-         auto usec = server.http_config->idle_timeout_us.load();
-         if (usec < 0)
-            return;
-         _expiration = steady_clock::now() + std::chrono::microseconds{usec};
-         _timer->expires_at(_expiration);
-         _timer->async_wait(
-             [this, self = std::move(self)](beast::error_code ec) mutable
-             {
-                if (ec || _closed)
-                {
-                   return;
-                }
-                if (steady_clock::now() >= _expiration)
-                {
-                   beast::error_code ec;
-                   close_impl(ec);
-                   if (ec)
-                   {
-                      PSIBASE_LOG(logger, warning) << "close: " << ec.message();
-                   }
-                   else
-                   {
-                      PSIBASE_LOG(logger, info) << "Idle connection closed";
-                   }
-                   _closed = true;
-                }
-             });
-      }
-
-      void stop_socket_timer()
-      {
-         _expiration = steady_clock::time_point::max();
-         _timer->cancel();
-      }
-
-      void on_read(beast::error_code ec, std::size_t bytes_transferred)
-      {
-         boost::ignore_unused(bytes_transferred);
-
-         if (_closed)
-            return;
-
-         // This means they closed the connection
-         if (ec == bhttp::error::end_of_stream)
-            return do_close();
-
-         if (ec)
-         {
-            fail(logger, ec, "read");
-            return close_on_error();
-         }
-
-         {
-            const auto& req = parser->get();
-            request_attrs.emplace(std::tuple{
-                boost::log::add_scoped_logger_attribute(
-                    logger, "RequestMethod",
-                    boost::log::attributes::constant{std::string(req.method_string())}),
-                boost::log::add_scoped_logger_attribute(
-                    logger, "RequestTarget",
-                    boost::log::attributes::constant{std::string(req.target())}),
-                boost::log::add_scoped_logger_attribute(
-                    logger, "RequestHost",
-                    boost::log::attributes::constant{std::string(req[bhttp::field::host])})});
-            PSIBASE_LOG(logger, debug) << "Received HTTP request";
-         }
-
-         // Send the response
-         handle_request(server, parser->release(), *this);
-
-         // The queue being empty means that we're waiting for the
-         // request to be processed asynchronously, and there are
-         // no pending writes.
-         if (is_empty())
-         {
-            assert(!can_read());
-            stop_socket_timer();
-         }
-
-         // If we aren't at the queue limit, try to pipeline another request
-         if (can_read())
-            do_read();
-      }
-
-      void do_close()
-      {
-         // Send a TCP shutdown
-         if (!_closed)
-         {
-            shutdown_impl();
-            _timer->cancel();  // cancel connection timer.
-            _closed = true;
-         }
-         // At this point the connection is closed gracefully
-      }
-      void close_on_error()
-      {
-         if (!_closed)
-         {
-            _timer->cancel();
-            beast::error_code ec;
-            close_impl(ec);
-            if (ec)
-            {
-               PSIBASE_LOG(logger, warning) << "close: " << ec.message();
-            }
-            _closed = true;
-         }
-      }
-
-      beast::flat_buffer                 buffer;
-      std::unique_ptr<net::steady_timer> _timer;
-      bool                               _closed = false;
-      steady_clock::time_point           _expiration;
-
-      // The parser is stored in an optional container so we can
-      // construct it from scratch it at the beginning of each new message.
-      boost::optional<bhttp::request_parser<bhttp::vector_body<char>>> parser;
-
-     public:
-      server_impl& server;
-
-      psibase::loggers::common_logger logger;
-      using scoped_attribute = decltype(boost::log::add_scoped_logger_attribute(
-          logger,
-          std::string(),
-          boost::log::attributes::constant{std::string()}));
-      std::optional<std::tuple<scoped_attribute, scoped_attribute, scoped_attribute>> request_attrs;
-
-     public:
-      bool pause_read = false;
-
-      // Returns `true` if we have reached the queue limit
-      bool is_full() const { return items.size() >= limit; }
-      bool can_read() const { return !is_full() && !pause_read; }
-      bool is_empty() const { return items.empty(); }
-
-      // Called when a message finishes sending
-      // Returns `true` if the caller should initiate a read
-      bool pop_queue()
-      {
-         BOOST_ASSERT(!items.empty());
-         const auto was_full = is_full();
-         items.erase(items.begin());
-         if (items.empty() && pause_read)
-            stop_socket_timer();
-         else
-            start_socket_timer();
-         if (!items.empty())
-            (*items.front())();
-         return was_full && !pause_read;
-      }
-
-      void on_write(bool close, beast::error_code ec, std::size_t bytes_transferred)
-      {
-         boost::ignore_unused(bytes_transferred);
-
-         if (_closed)
-            return;
-
-         if (ec)
-         {
-            fail(logger, ec, "write");
-            return close_on_error();
-         }
-
-         if (close)
-         {
-            // This means we should close the connection, usually because
-            // the response indicated the "Connection: close" semantic.
-            return do_close();
-         }
-
-         // Inform the queue that a write completed
-         if (pop_queue())
-         {
-            // Read another request
-            do_read();
-         }
-      }
-
-      // Called by the HTTP handler to send a response.
-      void operator()(message_type&& msg)
-      {
-         // This holds a work item
-         struct work_impl : work
-         {
-            http_session_base& self;
-            message_type       msg;
-
-            work_impl(http_session_base& self, message_type&& msg) : self(self), msg(std::move(msg))
-            {
-            }
-
-            void operator()() { self.write_response(std::move(msg)); }
-         };
-
-         {
-            BOOST_LOG_SCOPED_LOGGER_TAG(logger, "ResponseStatus",
-                                        static_cast<unsigned>(msg.result_int()));
-            BOOST_LOG_SCOPED_LOGGER_ATTR(logger, "ResponseBytes",
-                                         boost::log::attributes::constant<std::uint64_t>(
-                                             msg.payload_size() ? *msg.payload_size() : 0));
-            PSIBASE_LOG(logger, info) << "Handled HTTP request";
-            request_attrs.reset();
-         }
-
-         // Allocate and store the work
-         items.push_back(boost::make_unique<work_impl>(*this, std::move(msg)));
-
-         // If there was no previous work, start this one
-         if (items.size() == 1)
-         {
-            start_socket_timer();
-            (*items.front())();
-         }
-      }
-
-      void operator()(websocket_upgrade, request_type&& msg, accept_p2p_websocket_t f)
-      {
-         struct work_impl : work
-         {
-            http_session_base&     self;
-            request_type           request;
-            accept_p2p_websocket_t next;
-
-            work_impl(http_session_base& self, request_type&& msg, accept_p2p_websocket_t&& f)
-                : self(self), request(std::move(msg)), next(std::move(f))
-            {
-            }
-            void operator()() { self.accept_websocket(std::move(request), std::move(next)); }
-         };
-
-         {
-            BOOST_LOG_SCOPED_LOGGER_TAG(logger, "ResponseStatus",
-                                        static_cast<unsigned>(bhttp::status::switching_protocols));
-            PSIBASE_LOG(logger, info) << "Handled HTTP request";
-            request_attrs.reset();
-         }
-
-         // Allocate and store the work
-         items.push_back(boost::make_unique<work_impl>(*this, std::move(msg), std::move(f)));
-
-         // If there was no previous work, start this one
-         if (items.size() == 1)
-            (*items.front())();
-      }
-
-      virtual void write_response(message_type&& msg)                                      = 0;
-      virtual void accept_websocket(request_type&& request, accept_p2p_websocket_t&& next) = 0;
-      virtual void do_read()                                                               = 0;
-      virtual bool check_access(const authz_loopback&) const                               = 0;
-      virtual bool check_access(const authz_ip&) const                                     = 0;
-      virtual void post(std::function<void()>)                                             = 0;
-      virtual void close_impl(beast::error_code& ec)                                       = 0;
-      virtual void shutdown_impl()                                                         = 0;
-   };
-
-   template <typename T>
-   constexpr std::size_t function_arg_count = 0;
-
-   template <typename R, typename... T>
-   constexpr std::size_t function_arg_count<std::function<R(T...)>> = sizeof...(T);
-
-   beast::string_view remove_scheme(beast::string_view origin)
-   {
-      auto pos = origin.find("://");
-      if (pos != beast::string_view::npos)
-      {
-         return origin.substr(pos + 3);
-      }
-      return origin;
-   }
-
-   beast::string_view deport(beast::string_view host)
-   {
-      if (auto pos = host.rfind(':'); pos != std::string::npos)
-      {
-         if (host.find(']', pos) == std::string::npos)
-         {
-            return host.substr(0, pos);
-         }
-      }
-      return host;
-   }
-
-   bool is_admin(admin_none, const http_config&, beast::string_view)
-   {
-      return false;
-   }
-
-   bool is_admin(admin_any, const http_config&, beast::string_view)
-   {
-      return true;
-   }
-
-   bool is_admin(admin_any_native, const http_config& cfg, beast::string_view host)
-   {
-      return cfg.services.find(host) != cfg.services.end();
-   }
-
-   bool is_admin(AccountNumber admin, const http_config& cfg, beast::string_view host)
-   {
-      if (host.size() > cfg.host.size() && host.ends_with(cfg.host) &&
-          host[host.size() - cfg.host.size() - 1] == '.')
-      {
-         return admin.str() == host.substr(0, host.size() - cfg.host.size() - 1);
-      }
-      else
-      {
-         return false;
-      }
-   }
-
-   bool is_admin(const http_config& cfg, beast::string_view host)
-   {
-      host = deport(host);
-      return std::visit([&](const auto& admin) { return is_admin(admin, cfg, host); }, cfg.admin);
-   }
-
-   // Returns the access mode and an indication of whether the request contains
-   // an explicit authorization. The result access mode will not be greater than
-   // the argument.
-   std::pair<authz::mode_type, bool> get_access_impl(const auto&              req,
-                                                     const http_session_base& session,
-                                                     const auto&              auth,
-                                                     authz::mode_type         mode)
-   {
-      if constexpr (requires { session.check_access(auth); })
-      {
-         if (session.check_access(auth))
-         {
-            return {mode, false};
-         }
-         else
-         {
-            return {authz::none, false};
-         }
-      }
-      else
-      {
-         return {authz::none, false};
-      }
-   }
-   std::pair<authz::mode_type, bool> get_access_impl(const auto&              req,
-                                                     const http_session_base& session,
-                                                     const authz_any&         auth,
-                                                     authz::mode_type         mode)
-   {
-      return {mode, false};
-   }
-   std::pair<authz::mode_type, bool> get_access_impl(const auto&              req,
-                                                     const http_session_base& session,
-                                                     const authz_bearer&      auth,
-                                                     authz::mode_type         mode)
-   {
-      auto             authorization_b = req[bhttp::field::authorization];
-      std::string_view authorization{authorization_b.data(), authorization_b.size()};
-      std::string_view prefix = "Bearer ";
-      if (authorization.starts_with(prefix))
-      {
-         try
-         {
-            token_data token = decode_jwt(auth.key, authorization.substr(prefix.size()));
-            if (std::chrono::system_clock::now() >
-                std::chrono::system_clock::time_point{std::chrono::seconds(token.exp)})
-            {
-               return {authz::none, false};
-            }
-            return {static_cast<authz::mode_type>(mode & authz::mode_from_string(token.mode)),
-                    true};
-         }
-         catch (...)
-         {
-         }
-      }
-      return {authz::none, false};
-   }
-
-   std::pair<authz::mode_type, bool> get_access(server_impl&       server,
-                                                auto&              req,
-                                                http_session_base& session)
-   {
-      std::pair<authz::mode_type, bool> result{authz::none, false};
-      for (const authz& auth : server.http_config->admin_authz)
-      {
-         auto [mode, has_auth] = std::visit(
-             [&](const auto& a) { return get_access_impl(req, session, a, auth.mode); }, auth.data);
-         result.first = static_cast<authz::mode_type>(result.first | mode);
-         result.second |= has_auth;
-      }
-      return result;
-   }
-
-   // true: The client is authorized
-   // false: The client has a valid authorization, but it is insufficent to access this resource
-   // nullopt: The client does not have a valid authorization
-   std::optional<bool> check_access(server_impl&     server,
-                                    auto&            req,
-                                    auto&            send,
-                                    authz::mode_type required_mode)
-   {
-      auto [mode, has_auth] = get_access(server, req, send);
-      if ((mode & required_mode) == required_mode)
-      {
-         return true;
-      }
-      else if (has_auth)
-      {
-         return false;
-      }
-      else
-      {
-         return {};
-      }
-   }
-
-   std::string mkscope(authz::mode_type mode)
-   {
-      std::string result;
-      if (mode & authz::read)
-      {
-         result += "monitor";
-      }
-      else if (mode & authz::write)
-      {
-         if (!result.empty())
-            result.push_back(' ');
-         result += "admin";
-      }
-      return result;
-   }
-
-   std::pair<beast::string_view, beast::string_view> parse_uri(beast::string_view uri)
-   {
-      if (uri.starts_with("http://"))
-      {
-         uri = uri.substr(7);
-      }
-      else if (uri.starts_with("https://"))
-      {
-         uri = uri.substr(8);
-      }
-      else
-      {
-         return {};
-      }
-      auto pos         = uri.find('/');
-      auto authority   = uri.substr(0, pos);
-      auto path        = pos == std::string::npos ? "/" : uri.substr(pos);
-      auto enduserinfo = authority.find('@');
-      auto host = enduserinfo == std::string::npos ? authority : authority.substr(enduserinfo + 1);
-      return {host, path};
-   }
-
-   // Returns the host and path for the request. See RFC 9112 § 3.2 and 3.3
-   template <typename Body, typename Allocator>
-   std::pair<beast::string_view, beast::string_view> parse_request_target(
-       const bhttp::request<Body, bhttp::basic_fields<Allocator>>& request)
-   {
-      auto target = request.target();
-      if (target.starts_with('/') || target == "*")
-      {
-         return {request[bhttp::field::host], target};
-      }
-      else
-      {
-         // We are not a proxy, so we don't need to support authority-form
-         return parse_uri(target);
-      }
-   }
-
-   std::vector<char> to_vector(std::string_view s)
-   {
-      return std::vector(s.begin(), s.end());
-   }
-
-   struct QueryTimes
-   {
-      std::chrono::microseconds             packTime;
-      std::chrono::microseconds             serviceLoadTime;
-      std::chrono::microseconds             databaseTime;
-      std::chrono::microseconds             wasmExecTime;
-      std::chrono::steady_clock::time_point startTime;
-   };
-
-   template <typename F, typename E>
-   struct HttpSocket : AutoCloseSocket, std::enable_shared_from_this<HttpSocket<F, E>>
-   {
-      explicit HttpSocket(std::shared_ptr<http_session_base>&& session, F&& f, E&& err)
-          : session(std::move(session)), callback(std::forward<F>(f)), err(std::forward<E>(err))
-      {
-         this->session->pause_read = true;
-         this->once                = true;
-      }
-      void autoClose(const std::optional<std::string>& message) noexcept override
-      {
-         sendImpl([message = message.value_or("service did not send a response")](
-                      HttpSocket* self) mutable { return self->err(message); });
-      }
-      void send(std::span<const char> data) override
-      {
-         sendImpl([result = psio::from_frac<HttpReply>(data)](HttpSocket* self) mutable
-                  { return self->callback(std::move(result)); });
-      }
-      void sendImpl(auto make_reply)
-      {
-         auto self = this->shared_from_this();
-         session->post(
-             [self = std::move(self), make_reply = std::move(make_reply)]() mutable
-             {
-                auto session = std::move(self->session);
-                if (!session)
-                   return;
-                auto& trace         = self->trace;
-                auto& queryTimes    = self->queryTimes;
-                session->pause_read = false;
-                {
-                   auto  endTime = steady_clock::now();
-                   auto& logger  = session->logger;
-                   BOOST_LOG_SCOPED_LOGGER_TAG(logger, "Trace", std::move(trace));
-
-                   // TODO: consider bundling into a single attribute
-                   BOOST_LOG_SCOPED_LOGGER_TAG(logger, "PackTime", queryTimes.packTime);
-                   BOOST_LOG_SCOPED_LOGGER_TAG(logger, "ServiceLoadTime",
-                                               queryTimes.serviceLoadTime);
-                   BOOST_LOG_SCOPED_LOGGER_TAG(logger, "DatabaseTime", queryTimes.databaseTime);
-                   BOOST_LOG_SCOPED_LOGGER_TAG(logger, "WasmExecTime", queryTimes.wasmExecTime);
-                   BOOST_LOG_SCOPED_LOGGER_TAG(
-                       logger, "ResponseTime",
-                       std::chrono::duration_cast<std::chrono::microseconds>(endTime -
-                                                                             queryTimes.startTime));
-                   (*session)(make_reply(self.get()));
-                }
-                if (session->can_read())
-                   session->do_read();
-             });
-      }
-      std::shared_ptr<http_session_base> session;
-      F                                  callback;
-      E                                  err;
-      TransactionTrace                   trace;
-      QueryTimes                         queryTimes;
-   };
-
-   template <typename F, typename E>
-   auto makeHttpSocket(http_session_base& send, F&& f, E&& e)
-   {
-      return std::make_shared<HttpSocket<std::decay_t<F>, std::decay_t<E>>>(
-          send.shared_from_this(), std::forward<F>(f), std::forward<E>(e));
-   }
-
-   // This function produces an HTTP response for the given
-   // request. The type of the response object depends on the
-   // contents of the request, so the interface requires the
-   // caller to pass a generic lambda for receiving the response.
-   template <class Body, class Allocator>
-   void handle_request(server_impl&                                           server,
-                       bhttp::request<Body, bhttp::basic_fields<Allocator>>&& req,
-                       http_session_base&                                     send)
-   {
-      auto [req_host, req_target] = parse_request_target(req);
-      unsigned req_version        = req.version();
-      bool     req_keep_alive     = req.keep_alive();
-      bool     req_allow_cors     = !req_target.starts_with("/native/");
-
-      const auto set_cors = [&server, req_allow_cors](auto& res)
-      {
-         if (!server.http_config->allow_origin.empty() && req_allow_cors)
-         {
-            res.set(bhttp::field::access_control_allow_origin, server.http_config->allow_origin);
-            res.set(bhttp::field::access_control_allow_methods, "POST, GET, OPTIONS");
-            res.set(bhttp::field::access_control_allow_headers, "*");
-         }
-      };
-
-      const auto set_keep_alive = [&server, req_keep_alive](auto& res)
-      {
-         bool keep_alive = req_keep_alive && !server.http_config->status.load().shutdown;
-         res.keep_alive(keep_alive);
-         if (keep_alive)
-         {
-            if (auto usec = server.http_config->idle_timeout_us.load(); usec >= 0)
-            {
-               auto sec =
-                   std::chrono::duration_cast<std::chrono::seconds>(std::chrono::microseconds{usec})
-                       .count();
-               res.set(bhttp::field::keep_alive, "timeout=" + std::to_string(sec));
-            }
-         }
-      };
-
-      // Returns a method_not_allowed response
-      const auto method_not_allowed = [&server, set_cors, req_version, set_keep_alive](
-                                          beast::string_view target, beast::string_view method,
-                                          beast::string_view allowed_methods)
-      {
-         bhttp::response<bhttp::vector_body<char>> res{bhttp::status::method_not_allowed,
-                                                       req_version};
-         res.set(bhttp::field::server, BOOST_BEAST_VERSION_STRING);
-         res.set(bhttp::field::content_type, "text/html");
-         res.set(bhttp::field::allow, allowed_methods);
-         set_cors(res);
-         set_keep_alive(res);
-         res.body() = to_vector("The resource '" + std::string(target) +
-                                "' does not accept the method " + std::string(method) + ".");
-         res.prepare_payload();
-         return res;
-      };
-
-      // Returns an error response
-      const auto error =
-          [&server, set_cors, req_version, set_keep_alive](
-              bhttp::status status, beast::string_view why, const char* content_type = "text/html")
-      {
-         bhttp::response<bhttp::vector_body<char>> res{status, req_version};
-         res.set(bhttp::field::server, BOOST_BEAST_VERSION_STRING);
-         res.set(bhttp::field::content_type, content_type);
-         set_cors(res);
-         set_keep_alive(res);
-         res.body() = std::vector(why.begin(), why.end());
-         res.prepare_payload();
-         return res;
-      };
-
-      const auto not_found = [&error](beast::string_view target)
-      {
-         return error(bhttp::status::not_found,
-                      "The resource '" + std::string(target) + "' was not found.");
-      };
-      const auto bad_request = [&error](beast::string_view why)
-      { return error(bhttp::status::bad_request, why); };
-
-      // Returns an error response with a WWW-Authenticate header
-      const auto auth_error = [&server, set_cors, req_version, set_keep_alive](
-                                  bhttp::status status, std::string&& www_auth)
-      {
-         bhttp::response<bhttp::vector_body<char>> res{status, req_version};
-         res.set(bhttp::field::server, BOOST_BEAST_VERSION_STRING);
-         res.set(bhttp::field::content_type, "text/html");
-         res.set(bhttp::field::www_authenticate, std::move(www_auth));
-         set_cors(res);
-         set_keep_alive(res);
-         res.body() = to_vector("Not authorized");
-         res.prepare_payload();
-         return res;
-      };
-
-      const auto ok = [&server, set_cors, req_version, set_keep_alive](
-                          std::vector<char> reply, const char* content_type,
-                          const std::vector<HttpHeader>* headers = nullptr)
-      {
-         bhttp::response<bhttp::vector_body<char>> res{bhttp::status::ok, req_version};
-         res.set(bhttp::field::server, BOOST_BEAST_VERSION_STRING);
-         if (headers)
-            for (auto& h : *headers)
-               res.set(h.name, h.value);
-         res.set(bhttp::field::content_type, content_type);
-         set_cors(res);
-         set_keep_alive(res);
-         res.body() = std::move(reply);
-         res.prepare_payload();
-         return res;
-      };
-
-      const auto ok_no_content = [&server, set_cors, req_version, set_keep_alive]()
-      {
-         bhttp::response<bhttp::vector_body<char>> res{bhttp::status::ok, req_version};
-         res.set(bhttp::field::server, BOOST_BEAST_VERSION_STRING);
-         set_cors(res);
-         set_keep_alive(res);
-         res.prepare_payload();
-         return res;
-      };
-
-      const auto accepted = [&server, set_cors, req_version, set_keep_alive]()
-      {
-         bhttp::response<bhttp::vector_body<char>> res{bhttp::status::accepted, req_version};
-         res.set(bhttp::field::server, BOOST_BEAST_VERSION_STRING);
-         set_cors(res);
-         set_keep_alive(res);
-         res.prepare_payload();
-         return res;
-      };
-
-      // Returns true on success
-      // Sends an appropriate error response and returns false on failure
-      const auto check_admin_auth = [&server, &req, &send, &auth_error](authz::mode_type mode)
-      {
-         if (auto result = check_access(server, req, send, mode))
-         {
-            if (*result)
-               return true;
-            else
-            {
-               send(auth_error(bhttp::status::forbidden, "Bearer scope=\"" + mkscope(mode) + "\""));
-               return false;
-            }
-         }
-         else
-         {
-            send(auth_error(bhttp::status::unauthorized, "Bearer scope=\"" + mkscope(mode) + "\""));
-            return false;
-         }
-      };
-
-      // Used for requests that are not subject to CORS such as websockets
-      auto forbid_cross_origin = [&req, &req_host, &send, &error]()
-      {
-         if (auto iter = req.find(bhttp::field::origin); iter != req.end())
-         {
-            const auto& origin = iter->value();
-            if (origin == "null" || remove_scheme(origin) != req_host)
-            {
-               send(error(bhttp::status::bad_request, "Cross origin request refused"));
-               return true;
-            }
-         }
-         return false;
-      };
-
-      const auto run_native_handler = [&send, &req](auto&& request_handler, auto&& callback)
-      {
-         send.pause_read        = true;
-         auto post_back_to_http = [callback = static_cast<decltype(callback)>(callback),
-                                   session  = send.shared_from_this()](auto&& result) mutable
-         {
-            auto* p = session.get();
-            p->post(
-                [callback = std::move(callback), session = std::move(session),
-                 result = static_cast<decltype(result)>(result)]() mutable
-                {
-                   session->pause_read = false;
-                   (*session)(callback(std::move(result)));
-                   if (session->can_read())
-                      session->do_read();
-                });
-         };
-         if constexpr (function_arg_count<std::remove_cvref_t<decltype(request_handler)>> == 2)
-         {
-            request_handler(std::move(req.body()), std::move(post_back_to_http));
-         }
-         else
-         {
-            request_handler(std::move(post_back_to_http));
-         }
-      };
-
-      // The result should be a variant with a std::string representing an error
-      const auto& run_native_handler_json = [&](auto&& request_handler)
-      {
-         run_native_handler(
-             request_handler,
-             [error, ok](auto&& result)
-             {
-                return std::visit(
-                    [&](auto& body)
-                    {
-                       if constexpr (!std::is_same_v<std::decay_t<decltype(body)>, std::string>)
-                       {
-                          std::vector<char>   data;
-                          psio::vector_stream stream{data};
-                          psio::to_json(body, stream);
-                          return ok(std::move(data), "application/json");
-                       }
-                       else
-                       {
-                          return error(bhttp::status::internal_server_error, body);
-                       }
-                    },
-                    result);
-             });
-      };
-
-      // Anything that can serialized as json
-      const auto& run_native_handler_json_nofail = [&](auto&& request_handler)
-      {
-         run_native_handler(request_handler,
-                            [ok](auto&& result)
-                            {
-                               std::vector<char>   data;
-                               psio::vector_stream stream{data};
-                               psio::to_json(result, stream);
-                               return ok(std::move(data), "application/json");
-                            });
-      };
-
-      // optional<string>
-      auto run_native_handler_no_content = [&](auto&& request_handler)
-      {
-         run_native_handler(request_handler,
-                            [ok_no_content, error](auto&& result)
-                            {
-                               if (result)
-                               {
-                                  return error(bhttp::status::internal_server_error, *result);
-                               }
-                               else
-                               {
-                                  return ok_no_content();
-                               }
-                            });
-      };
-
-      // variant<string, function<vector<char>()>>
-      auto run_native_handler_generic = [&](auto& request_handler, const char* content_type)
-      {
-         run_native_handler(
-             request_handler,
-             [error, ok, content_type](auto&& result)
-             {
-                return std::visit(
-                    [&](auto& body)
-                    {
-                       if constexpr (!std::is_same_v<std::decay_t<decltype(body)>, std::string>)
-                       {
-                          return ok(body(), content_type);
-                       }
-                       else
-                       {
-                          return error(bhttp::status::internal_server_error, body);
-                       }
-                    },
-                    result);
-             });
-      };
-
-      auto run_native_handler_generic_nofail = [&](auto& request_handler, const char* content_type)
-      {
-         run_native_handler(request_handler, [ok, content_type](auto&& result)
-                            { return ok(result(), content_type); });
-      };
-
-      try
-      {
-         std::shared_lock l{server.http_config->mutex};
-         if (req_target == "")
-         {
-            return send(bad_request("Invalid request target"));
-         }
-         else if (!req_target.starts_with("/native"))
-         {
-            auto host = deport(req_host);
-
-            if (auto iter = server.http_config->services.find(host);
-                iter != server.http_config->services.end())
-            {
-               auto file = iter->second.find(req_target);
-               if (file != iter->second.end())
-               {
-                  if (req.method() == bhttp::verb::options)
-                  {
-                     return send(ok_no_content());
-                  }
-                  else if (req.method() != bhttp::verb::get)
-                  {
-                     return send(
-                         method_not_allowed(req.target(), req.method_string(), "GET, OPTIONS"));
-                  }
-
-                  std::vector<char> contents;
-                  // read file
-                  auto          size = std::filesystem::file_size(file->second.path);
-                  std::ifstream in(file->second.path, std::ios_base::binary);
-                  l.unlock();
-                  contents.resize(size);
-                  in.read(contents.data(), contents.size());
-                  return send(ok(std::move(contents), file->second.content_type.c_str()));
-               }
-               else
-               {
-                  return send(not_found(req.target()));
-               }
-            }
-
-            if (server.http_config->host.empty())
-               return send(not_found(req.target()));
-
-            if (req.method() == bhttp::verb::options)
-            {
-               return send(ok_no_content());
-            }
-
-            auto        startTime = steady_clock::now();
-            HttpRequest data;
-            if (req.method() == bhttp::verb::get)
-               data.method = "GET";
-            else if (req.method() == bhttp::verb::post)
-               data.method = "POST";
-            else
-               return send(
-                   method_not_allowed(req.target(), req.method_string(), "GET, POST, OPTIONS"));
-            data.host        = {host.begin(), host.size()};
-            data.rootHost    = server.http_config->host;
-            data.target      = std::string(req_target);
-            data.contentType = (std::string)req[bhttp::field::content_type];
-            data.body        = std::move(req.body());
-
-            // Do not use any reconfigurable members of server.http_config after this point
-            l.unlock();
-
-            // TODO: time limit
-            auto system = server.sharedState->getSystemContext();
-
-            psio::finally f{[&]() { server.sharedState->addSystemContext(std::move(system)); }};
-            BlockContext  bc{*system, system->sharedDatabase.getHead(),
-                            system->sharedDatabase.createWriter(), true};
-            bc.start();
-            if (bc.needGenesisAction)
-               return send(error(bhttp::status::internal_server_error,
-                                 "Need genesis block; use 'psibase boot' to boot chain"));
-
-            SignedTransaction  trx;
-            TransactionTrace   trace;
-            TransactionContext tc{bc, trx, trace, true, false, true};
-            ActionTrace&       atrace        = trace.actionTraces.emplace_back();
-            auto               startExecTime = steady_clock::now();
-
-            auto socket = makeHttpSocket(
-                send,
-                [req_version, set_cors, set_keep_alive](HttpReply&& reply)
-                {
-                   bhttp::response<bhttp::vector_body<char>> res{
-                       bhttp::int_to_status(static_cast<std::uint16_t>(reply.status)), req_version};
-                   res.set(bhttp::field::server, BOOST_BEAST_VERSION_STRING);
-                   for (auto& h : reply.headers)
-                      res.set(h.name, h.value);
-                   res.set(bhttp::field::content_type, reply.contentType);
-                   set_cors(res);
-                   set_keep_alive(res);
-                   res.body() = std::move(reply.body);
-                   res.prepare_payload();
-                   return res;
-                },
-                [error](const std::string& message)
-                { return error(bhttp::status::internal_server_error, message); });
-            system->sockets->add(socket, &tc.ownedSockets);
-
-            auto setStatus = psio::finally(
-                [&]
-                {
-                   auto endExecTime = steady_clock::now();
-
-                   socket->trace = std::move(trace);
-
-                   socket->queryTimes.packTime =
-                       std::chrono::duration_cast<std::chrono::microseconds>(startExecTime -
-                                                                             startTime);
-                   socket->queryTimes.serviceLoadTime =
-                       std::chrono::duration_cast<std::chrono::microseconds>(
-                           endExecTime - startExecTime - tc.getBillableTime());
-                   socket->queryTimes.databaseTime =
-                       std::chrono::duration_cast<std::chrono::microseconds>(tc.databaseTime);
-                   socket->queryTimes.wasmExecTime =
-                       std::chrono::duration_cast<std::chrono::microseconds>(tc.getBillableTime() -
-                                                                             tc.databaseTime);
-                   socket->queryTimes.startTime = startTime;
-                });
-
-            try
-            {
-               Action action{
-                   .sender  = AccountNumber(),
-                   .service = proxyServiceNum,
-                   .rawData = psio::convert_to_frac(std::tuple(socket->id, std::move(data))),
-               };
-               tc.execServe(action, atrace);
-            }
-            catch (...)
-            {
-               // An error will be sent by the transaction context, if needed
-            }
-            if (!tc.ownedSockets.owns(*system->sockets, *socket))
-            {
-               auto  error  = trace.error;
-               auto& logger = send.logger;
-               BOOST_LOG_SCOPED_LOGGER_TAG(logger, "Trace", std::move(trace));
-               if (error)
-                  PSIBASE_LOG(logger, warning)
-                      << proxyServiceNum.str() << "::serve failed: " << *error;
-               else
-                  PSIBASE_LOG(logger, info) << proxyServiceNum.str() << "::serve succeeded";
-            }
-         }  // !native
-         else if (req_target == "/native/push_boot" && server.http_config->push_boot_async)
-         {
-            if (!server.http_config->enable_transactions)
-               return send(not_found(req.target()));
-
-            if (req.method() != bhttp::verb::post)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "POST"));
-            }
-
-            if (auto content_type = req.find(bhttp::field::content_type); content_type != req.end())
-            {
-               if (content_type->value() != "application/octet-stream")
-               {
-                  return send(error(bhttp::status::unsupported_media_type,
-                                    "Content-Type must be application/octet-stream\n"));
-               }
-            }
-
-            if (forbid_cross_origin())
-            {
-               return;
-            }
-
-            run_native_handler_json(server.http_config->push_boot_async);
-         }  // push_boot
-         else if (req_target == "/native/push_transaction" &&
-                  server.http_config->push_transaction_async)
-         {
-            if (!server.http_config->enable_transactions)
-               return send(not_found(req.target()));
-
-            if (req.method() != bhttp::verb::post)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "POST"));
-            }
-
-            if (auto content_type = req.find(bhttp::field::content_type); content_type != req.end())
-            {
-               if (content_type->value() != "application/octet-stream")
-               {
-                  return send(error(bhttp::status::unsupported_media_type,
-                                    "Content-Type must be application/octet-stream\n"));
-               }
-            }
-
-            if (forbid_cross_origin())
-            {
-               return;
-            }
-
-            run_native_handler_json(server.http_config->push_transaction_async);
-         }  // push_transaction
-         else if (req_target == "/native/p2p" && websocket::is_upgrade(req) &&
-                  !boost::type_erasure::is_empty(server.http_config->accept_p2p_websocket) &&
-                  server.http_config->enable_p2p)
-         {
-            if (forbid_cross_origin())
-               return;
-            // Stop reading HTTP requests
-            send.pause_read = true;
-            send(websocket_upgrade{}, std::move(req), server.http_config->accept_p2p_websocket);
-            return;
-         }
-         else if (req_target == "/native/admin/status")
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() == bhttp::verb::get)
-            {
-               if (!check_admin_auth(authz::mode_type::read))
-               {
-                  return;
-               }
-               auto status = server.http_config->status.load();
-               if (status.needgenesis)
-               {
-                  if (!server.sharedState->needGenesis())
-                  {
-                     status = atomic_set_field(server.http_config->status,
-                                               [](auto& tmp) { tmp.needgenesis = false; });
-                  }
-               }
-               std::vector<char>   body;
-               psio::vector_stream stream{body};
-               to_json(status, stream);
-               send(ok(std::move(body), "application/json"));
-            }
-            else
-            {
-               send(method_not_allowed(req.target(), req.method_string(), "GET"));
-            }
-         }
-         else if (req_target == "/native/admin/shutdown")
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::post)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "POST"));
-            }
-            if (!check_admin_auth(authz::mode_type::write))
-            {
-               return;
-            }
-            if (req[bhttp::field::content_type] != "application/json")
-            {
-               return send(error(bhttp::status::unsupported_media_type,
-                                 "Content-Type must be application/json\n"));
-            }
-            server.http_config->shutdown(std::move(req.body()));
-            return send(accepted());
-         }
-         else if (req_target == "/native/admin/perf" && server.http_config->get_perf)
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::get)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "GET"));
-            }
-            if (!check_admin_auth(authz::mode_type::read))
-            {
-               return;
-            }
-            run_native_handler_generic_nofail(server.http_config->get_perf, "application/json");
-         }
-         else if (req_target == "/native/admin/metrics" && server.http_config->get_metrics)
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::get)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "GET"));
-            }
-            if (!check_admin_auth(authz::mode_type::read))
-            {
-               return;
-            }
-            run_native_handler_generic_nofail(
-                server.http_config->get_metrics,
-                "application/openmetrics-text; version=1.0.0; charset=utf-8");
-         }
-         else if (req_target == "/native/admin/peers" && server.http_config->get_peers)
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::get)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "GET"));
-            }
-            if (!check_admin_auth(authz::mode_type::read))
-            {
-               return;
-            }
-
-            // returns json list of {id:int,endpoint:string}
-            run_native_handler_json_nofail(server.http_config->get_peers);
-         }
-         else if (req_target == "/native/admin/connect" && server.http_config->connect)
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::post)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "POST"));
-            }
-            if (!check_admin_auth(authz::mode_type::write))
-            {
-               return;
-            }
-
-            if (req[bhttp::field::content_type] != "application/json")
-            {
-               send(error(bhttp::status::unsupported_media_type,
-                          "Content-Type must be application/json\n"));
-               return;
-            }
-
-            run_native_handler_no_content(server.http_config->connect);
-         }
-         else if (req_target == "/native/admin/disconnect" && server.http_config->disconnect)
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::post)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "POST"));
-            }
-            if (!check_admin_auth(authz::mode_type::write))
-            {
-               return;
-            }
-
-            if (req[bhttp::field::content_type] != "application/json")
-            {
-               send(error(bhttp::status::unsupported_media_type,
-                          "Content-Type must be application/json\n"));
-               return;
-            }
-
-            run_native_handler_no_content(server.http_config->disconnect);
-         }
-         else if (req_target == "/native/admin/log" && websocket::is_upgrade(req))
-         {
-            if (!is_admin(*server.http_config, req_host))
-               return send(not_found(req.target()));
-            if (forbid_cross_origin())
-               return;
-            if (!check_admin_auth(authz::mode_type::read))
-            {
-               return;
-            }
-            send.pause_read = true;
-            send(websocket_upgrade{}, std::move(req),
-                 [&server](auto&& stream)
-                 {
-                    using stream_type = std::decay_t<decltype(stream)>;
-                    auto session =
-                        std::make_shared<websocket_log_session<stream_type>>(std::move(stream));
-                    server.register_connection(session);
-                    websocket_log_session<stream_type>::run(std::move(session));
-                 });
-         }
-         else if (req_target == "/native/admin/config")
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() == bhttp::verb::get)
-            {
-               if (!check_admin_auth(authz::mode_type::read_write))
-               {
-                  return;
-               }
-               run_native_handler_generic_nofail(server.http_config->get_config,
-                                                 "application/json");
-            }
-            else if (req.method() == bhttp::verb::put)
-            {
-               if (!check_admin_auth(authz::mode_type::write))
-               {
-                  return;
-               }
-               if (req[bhttp::field::content_type] != "application/json")
-               {
-                  return send(error(bhttp::status::unsupported_media_type,
-                                    "Content-Type must be application/json\n"));
-               }
-               run_native_handler_no_content(server.http_config->set_config);
-            }
-            else
-            {
-               send(method_not_allowed(req.target(), req.method_string(), "GET, PUT"));
-            }
-            return;
-         }
-         else if (req_target == "/native/admin/keys")
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() == bhttp::verb::get)
-            {
-               if (!check_admin_auth(authz::mode_type::read))
-               {
-                  return;
-               }
-               run_native_handler_generic_nofail(server.http_config->get_keys, "application/json");
-            }
-            else if (req.method() == bhttp::verb::post)
-            {
-               if (!check_admin_auth(authz::mode_type::write))
-               {
-                  return;
-               }
-               if (req[bhttp::field::content_type] != "application/json")
-               {
-                  return send(error(bhttp::status::unsupported_media_type,
-                                    "Content-Type must be application/json\n"));
-               }
-
-               run_native_handler_generic(server.http_config->new_key, "application/json");
-            }
-            else
-            {
-               send(method_not_allowed(req.target(), req.method_string(), "GET, POST"));
-            }
-            return;
-         }
-         else if (req_target == "/native/admin/keys/devices")
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::get)
-            {
-               send(method_not_allowed(req.target(), req.method_string(), "GET"));
-            }
-            if (!check_admin_auth(authz::mode_type::read))
-            {
-               return;
-            }
-            run_native_handler_generic(server.http_config->get_pkcs11_tokens, "application/json");
-         }
-         else if (req_target == "/native/admin/keys/unlock")
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::post)
-            {
-               send(method_not_allowed(req.target(), req.method_string(), "POST"));
-            }
-            if (!check_admin_auth(authz::mode_type::write))
-            {
-               return;
-            }
-            if (req[bhttp::field::content_type] != "application/json")
-            {
-               return send(error(bhttp::status::unsupported_media_type,
-                                 "Content-Type must be application/json\n"));
-            }
-            run_native_handler_no_content(server.http_config->unlock_keyring);
-         }
-         else if (req_target == "/native/admin/keys/lock")
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::post)
-            {
-               send(method_not_allowed(req.target(), req.method_string(), "POST"));
-            }
-            if (!check_admin_auth(authz::mode_type::write))
-            {
-               return;
-            }
-            if (req[bhttp::field::content_type] != "application/json")
-            {
-               return send(error(bhttp::status::unsupported_media_type,
-                                 "Content-Type must be application/json\n"));
-            }
-            run_native_handler_no_content(server.http_config->lock_keyring);
-         }
-         else if (req_target == "/native/admin/login")
-         {
-            if (!is_admin(*server.http_config, req_host))
-            {
-               return send(not_found(req.target()));
-            }
-            if (req.method() != bhttp::verb::post)
-            {
-               return send(method_not_allowed(req.target(), req.method_string(), "POST"));
-            }
-
-            if (req[bhttp::field::content_type] != "application/json")
-            {
-               send(error(bhttp::status::unsupported_media_type,
-                          "Content-Type must be application/json\n"));
-               return;
-            }
-
-            const token_key* key = nullptr;
-            for (const auto& auth : server.http_config->admin_authz)
-            {
-               if (const auto* bearer = std::get_if<authz_bearer>(&auth.data))
-               {
-                  key = &bearer->key;
-               }
-            }
-            if (!key)
-            {
-               return send(error(bhttp::status::internal_server_error,
-                                 "Login requires bearer tokens to be enabled"));
-            }
-            auto [mode, has_auth] = get_access(server, req, send);
-
-            auto input = std::move(req.body());
-            input.push_back('\0');
-            input.push_back('\0');
-            input.push_back('\0');
-            psio::json_token_stream stream{input.data()};
-            auto                    params = psio::from_json<token_data>(stream);
-            if (!params.exp)
-            {
-               params.exp = std::chrono::time_point_cast<std::chrono::seconds>(
-                                std::chrono::system_clock::now() + std::chrono::hours(1))
-                                .time_since_epoch()
-                                .count();
-            }
-            if (params.mode.empty())
-            {
-               if (mode == authz::read)
-               {
-                  params.mode = "r";
-               }
-               else
-               {
-                  params.mode = "rw";
-               }
-            }
-            auto required_mode = authz::mode_from_string(params.mode);
-            if ((required_mode & mode) != required_mode)
-            {
-               if (has_auth)
-                  return send(auth_error(bhttp::status::forbidden,
-                                         "Bearer scope=\"" + mkscope(required_mode) + "\""));
-               else
-                  return send(auth_error(bhttp::status::unauthorized,
-                                         "Bearer scope=\"" + mkscope(required_mode) + "\""));
-            }
-            auto token = encode_jwt(*key, params);
-
-            std::vector<char>   data;
-            psio::vector_stream out{data};
-            psio::to_json(
-                token_response{.accessToken = token, .exp = params.exp, .mode = params.mode}, out);
-            return send(ok(std::move(data), "application/json"));
-         }
-         else
-         {
-            return send(not_found(req.target()));
-         }
-      }
-      catch (const std::exception& e)
-      {
-         return send(
-             error(bhttp::status::internal_server_error, "exception: " + std::string(e.what())));
-      }
-      catch (...)
-      {
-         // TODO: elog("query failed: unknown exception");
-         return send(
-             error(bhttp::status::internal_server_error, "query failed: unknown exception\n"));
-      }
-   }  // handle_request
 
    // Handles an HTTP server connection
    template <typename SessionType>
